Initial work for the com.d567.app API
Worked on the com.d567.app package - the front-end of the D567 API. This
includes Application Settings and the Trace Functions that a client app
will use. This required some re-working of the other packages to take
advantage of these settings.

This broke the ContentProvider example project since there is currently
no standard way to communicate the Application Settings to the another
application so that it can use the ContentProvider. This will be my next
focus.