Added Package List Request
When the Package List Request is broadcasted, each application setup to
handle that broadcast will in turn add its package name - a unique
identifier of an application - to the return results. The caller can use
that list of packages to then query more detailed information from the
individual applications - like requesting its Application Settings.

The example applications have been updated to use this feature - whereas
before the package names had to be hard-coded.