Integration + Transition Example
The D567 APIs integration with the Android OS has been improved. The
com.d567.app.Application class now derives from android.os.Applicaiton
and can be specified as the Application class to be used via the
Manifest. In addition, the Application Settings are now read from the
Application's Resources.

A new example program has been which tests/demonstrates the D567 APIs
ability to persist across activities and throughout the lifetime of an
application.

The older examples have been updated to work with the changes to the API