Added an Unhandled Exception Handler + Example
The Application will now set itself up as the Unhandled Exception
Handler for the thread in which it is created (default: the main UI
thread). When this occurs, the Application will Trace the error to the
current session (if applicable) and then send a request to the D567 App
to display the error message.

The ErrorHandlerExample Project has been added to demonstrate this
functionality - albeit the D567 App is not at a place where it can
handle this functionality yet.