Added API support for Save States
Added the Database + Request/Receiver Code for supporting Save States.
However, the D567 App is not a point yet where this feature can be
utilized/tested - so there is no example program. A Content Provider
will need to be created to make the save states available to the D567
App as well.